requirements.txt
Babel==2.7.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Django==2.2.1
django-cors-headers==3.2.1
django-jalali==4.0.0
django-phonenumber-field==3.0.1
djangorestframework==3.9.4
furl==2.0.0
idna==2.8
jdatetime==3.6.2
kavenegar==1.1.2
orderedmultidict==1.0.1
phonenumbers==8.10.13
Pillow==6.0.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
urllib3==1.25.3
